Refactored schemas/User to schemas/user removing schemas/user.py
