+++ frontend/src/__tests__/UserList.test.js
@@ -32,7 +32,7 @@ describe('<UserList />', () => {
                     userName: 'testuser@testemail.gc.ca',
                     admin: true,
                     tfa: false,
-                    displayName: 'Test User Esq.',
+                    displayName: 'Test user Esq.',
                   },
                 },
               ],
@@ -92,7 +92,7 @@ describe('<UserList />', () => {
                     userName: 'testuser@testemail.gc.ca',
                     admin: true,
                     tfa: false,
-                    displayName: 'Test User Esq.',
+                    displayName: 'Test user Esq.',
                   },
                 },
               ],
