+++ frontend/src/locales/fr.po
@@ -97,7 +97,7 @@ msgid "Create Account"
 msgstr "Créer un compte"
 
 #: src/UserList.js:58
-msgid "Create User"
+msgid "Create user"
 msgstr "Créer un utilisateur"
 
 #: src/CreateUserPage.js:103
